Use react-router to navigate in touchbar 🤯
