+++ src/touchbar/index.js
@@ -4,4 +4,5 @@
 
 
 export { default as TouchBar } from './TouchBar';
+export { default as navigateByActivePane } from './navigateByActivePane';
 export * from './touchBarByActivePane';
